Factor out definitions to be local to editor
