+++ browser/src/Services/Language/index.ts
@@ -1,6 +1,5 @@
 export * from "./addInsertModeLanguageFunctionality"
 export * from "./CodeAction"
-export * from "./Definition"
 export * from "./DefinitionRequestor"
 export * from "./Edits"
 export * from "./FindAllReferences"
