+++ browser/src/UI/containers/DefinitionContainer.ts
@@ -3,7 +3,6 @@ import { connect } from "react-redux"
 import * as types from "vscode-languageserver-types"
 
 import * as Selectors from "./../Selectors"
-import { getActiveDefinition } from "./../selectors/DefinitionSelectors"
 
 import * as State from "./../../Editor/NeovimEditor/NeovimEditorStore"
 
@@ -13,6 +12,8 @@ const emptyRange = types.Range.create(
     types.Position.create(-1, -1),
     types.Position.create(-1, -1),
 )
+
+const getActiveDefinition = (state: State.IState) => state.definition
 
 const mapStateToProps = (state: State.IState): IDefinitionProps => {
